fixed typings
