+++ components/LoginForm.tsx
@@ -9,7 +9,7 @@ import { useAuth } from '@/context/auth';
 import { ErrorText } from '@/components/ErrorText';
 import { Link } from '@/components/Link';
 import { setErrorsFromGraphQLErrors } from '@/utils/setErrors';
-import { LoginMutationVariables, useLoginMutation } from '@/types';
+import { LoginMutationVariables, useLoginMutation } from '@/typesgql';
 
 export const LOGIN_MUTATION = gql`
   mutation login($email: String!, $password: String!) {
