+++ prisma/scripts/exampleUserScript.ts
@@ -1,7 +1,7 @@
 import { Prisma } from '@prisma/client';
 
 import { hashPassword } from '@/services/auth';
-import { Role } from '@/types';
+import { Role } from '@/typesgql';
 import { seedUsers } from '@/prisma/seeds/users';
 import { prisma } from '@/lib/prisma';
