+++ prisma/seeds/users/prismaRunner.ts
@@ -1,7 +1,7 @@
 import { Prisma } from '@prisma/client';
 
 import { prisma } from '@/lib/prisma';
-import { User } from '@/types';
+import { User } from '@/typesgql';
 
 type SeedUserResult = Pick<User, 'id' | 'email'>;
