+++ tsconfig.json
@@ -1,11 +1,7 @@
 {
   "compilerOptions": {
     "target": "es2017",
-    "lib": [
-      "dom",
-      "dom.iterable",
-      "esnext"
-    ],
+    "lib": ["dom", "dom.iterable", "esnext"],
     "allowJs": true,
     "skipLibCheck": true,
     "strict": true,
@@ -24,10 +20,7 @@
         "name": "nexus/typescript-language-service"
       }
     ],
-    "typeRoots": [
-      "node_modules/@types",
-      "types"
-    ],
+    "typeRoots": ["node_modules/@types", "types"],
     "rootDir": ".",
     "baseUrl": ".",
     "paths": {
@@ -35,16 +28,6 @@
     },
     "incremental": true
   },
-  "include": [
-    ".",
-    "next-env.d.ts",
-    "types.ts",
-    "types.d.ts",
-    "**/*.ts",
-    "**/*.tsx"
-  ],
-  "exclude": [
-    "node_modules",
-    "cypress"
-  ]
+  "include": [".", "next-env.d.ts", "typesgql.ts", "types.d.ts", "**/*.ts", "**/*.tsx"],
+  "exclude": ["node_modules", "cypress"]
 }
