+++ types.ts
@@ -1,329 +1,72 @@
-/* eslint-disable */
-import { gql } from '@apollo/client';
-import * as Apollo from '@apollo/client';
-export type Maybe<T> = T | null;
-export type InputMaybe<T> = Maybe<T>;
-export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
-export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
-export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
-const defaultOptions = {} as const;
-/** All built-in and custom scalars, mapped to their actual values */
-export type Scalars = {
-  ID: string;
-  String: string;
-  Boolean: boolean;
-  Int: number;
-  Float: number;
-  /** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
-  DateTime: any;
-  /** A field whose value conforms to the standard internet email address format as specified in HTML Spec: https://html.spec.whatwg.org/multipage/input.html#valid-e-mail-address. */
-  EmailAddress: any;
-  /** The `JSONObject` scalar type represents JSON objects as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
-  JSONObject: any;
-  /** A field whose value conforms to the standard E.164 format as specified in: https://en.wikipedia.org/wiki/E.164. Basically this is +17895551234. */
-  PhoneNumber: any;
-  /** A field whose value conforms to the standard URL format as specified in RFC3986: https://www.ietf.org/rfc/rfc3986.txt. */
-  URL: any;
-};
+import Stripe from 'stripe';
 
-/** Payload returned if login or signup is successful */
-export type AuthPayload = {
-  __typename?: 'AuthPayload';
-  /** The current JWT token. Use in Authentication header */
-  token?: Maybe<Scalars['String']>;
-  /** The logged in user */
-  user?: Maybe<User>;
-};
-
-/** A way to filter date fields. Meant to pass to prisma where clause */
-export type DateFilter = {
-  equals?: InputMaybe<Scalars['DateTime']>;
-  gt?: InputMaybe<Scalars['DateTime']>;
-  gte?: InputMaybe<Scalars['DateTime']>;
-  in?: InputMaybe<Array<Scalars['DateTime']>>;
-  lt?: InputMaybe<Scalars['DateTime']>;
-  lte?: InputMaybe<Scalars['DateTime']>;
-  notIn?: InputMaybe<Array<Scalars['DateTime']>>;
-};
-
-export type Mutation = {
-  __typename?: 'Mutation';
-  /** Create User for an account */
-  createUser?: Maybe<User>;
-  /** Login to an existing account */
-  login?: Maybe<AuthPayload>;
-  /** Signup for an account */
-  signup?: Maybe<AuthPayload>;
-};
-
-
-export type MutationCreateUserArgs = {
-  data: UserCreateInput;
-};
-
-
-export type MutationLoginArgs = {
-  email: Scalars['String'];
-  password: Scalars['String'];
-};
-
-
-export type MutationSignupArgs = {
-  data: SignupInput;
-};
-
-/** A way to filter number fields. Meant to pass to prisma where clause */
-export type NumberFilter = {
-  equals?: InputMaybe<Scalars['Int']>;
-  gt?: InputMaybe<Scalars['Int']>;
-  gte?: InputMaybe<Scalars['Int']>;
-  in?: InputMaybe<Array<Scalars['Int']>>;
-  lt?: InputMaybe<Scalars['Int']>;
-  lte?: InputMaybe<Scalars['Int']>;
-  notIn?: InputMaybe<Array<Scalars['Int']>>;
-};
-
-/** A User Profile */
-export type Profile = {
-  __typename?: 'Profile';
-  createdAt: Scalars['DateTime'];
-  firstName: Scalars['String'];
-  /** The first and last name of a user */
-  fullName?: Maybe<Scalars['String']>;
-  id: Scalars['ID'];
-  lastName: Scalars['String'];
-  updatedAt: Scalars['DateTime'];
-  user: User;
-};
-
-/** Profile Input for relational Create */
-export type ProfileCreateInput = {
-  firstName: Scalars['String'];
-  lastName: Scalars['String'];
-};
-
-/** Input to Add a new user */
-export type ProfileRelationalCreateInput = {
-  create: ProfileCreateInput;
-};
-
-export type Query = {
-  __typename?: 'Query';
-  /** Returns the currently logged in user */
-  me?: Maybe<User>;
-  user?: Maybe<User>;
-  users?: Maybe<Array<Maybe<User>>>;
-};
-
-
-export type QueryUserArgs = {
-  where: UserWhereUniqueInput;
-};
-
-export enum Role {
-  ADMIN = 'ADMIN',
-  USER = 'USER'
+export interface Song {
+  id: string;
+  user_id: string;
+  author: string;
+  title: string;
+  song_path: string;
+  image_path: string;
 }
 
-/** Input required for a user to signup */
-export type SignupInput = {
-  email: Scalars['String'];
-  password: Scalars['String'];
-  profile: SignupProfileInput;
-};
-
-/** Input required for Profile Create on Signup. */
-export type SignupProfileCreateInput = {
-  firstName: Scalars['String'];
-  lastName: Scalars['String'];
-};
-
-/** Input required for Profile on Signup. */
-export type SignupProfileInput = {
-  create: SignupProfileCreateInput;
-};
-
-/** Sort direction for filtering queries (ascending or descending) */
-export enum SortOrder {
-  ASC = 'asc',
-  DESC = 'desc'
+export interface Product {
+  id: string;
+  active?: boolean;
+  name?: string;
+  description?: string;
+  image?: string;
+  metadata?: Stripe.Metadata;
 }
 
-/** A way to filter string fields. Meant to pass to prisma where clause */
-export type StringFilter = {
-  contains?: InputMaybe<Scalars['String']>;
-  endsWith?: InputMaybe<Scalars['String']>;
-  equals?: InputMaybe<Scalars['String']>;
-  gt?: InputMaybe<Scalars['String']>;
-  gte?: InputMaybe<Scalars['String']>;
-  in?: InputMaybe<Array<Scalars['String']>>;
-  lt?: InputMaybe<Scalars['String']>;
-  lte?: InputMaybe<Scalars['String']>;
-  notIn?: InputMaybe<Array<Scalars['String']>>;
-  startsWith?: InputMaybe<Scalars['String']>;
-};
-
-/** A User */
-export type User = {
-  __typename?: 'User';
-  createdAt: Scalars['DateTime'];
-  email?: Maybe<Scalars['String']>;
-  id: Scalars['ID'];
-  profile?: Maybe<Profile>;
-  roles: Array<Role>;
-  updatedAt: Scalars['DateTime'];
-};
-
-/** Input to Add a new user */
-export type UserCreateInput = {
-  email: Scalars['String'];
-  password: Scalars['String'];
-  profile?: InputMaybe<ProfileRelationalCreateInput>;
-  roles?: InputMaybe<Array<InputMaybe<Role>>>;
-};
-
-/** Order users by a specific field */
-export type UserOrderByInput = {
-  createdAt?: InputMaybe<SortOrder>;
-  email?: InputMaybe<SortOrder>;
-  updatedAt?: InputMaybe<SortOrder>;
-};
-
-/** Input to find users based other fields */
-export type UserWhereInput = {
-  email?: InputMaybe<StringFilter>;
-  id?: InputMaybe<Scalars['Int']>;
-};
-
-/** Input to find users based on unique fields */
-export type UserWhereUniqueInput = {
-  email?: InputMaybe<Scalars['String']>;
-  id?: InputMaybe<Scalars['ID']>;
-};
-
-export type LoginMutationVariables = Exact<{
-  email: Scalars['String'];
-  password: Scalars['String'];
-}>;
-
-
-export type LoginMutation = { __typename?: 'Mutation', login?: { __typename?: 'AuthPayload', token?: string | null } | null };
-
-export type SignupMutationVariables = Exact<{
-  data: SignupInput;
-}>;
-
-
-export type SignupMutation = { __typename?: 'Mutation', signup?: { __typename?: 'AuthPayload', token?: string | null, user?: { __typename?: 'User', id: string } | null } | null };
-
-export type MeQueryVariables = Exact<{ [key: string]: never; }>;
-
-
-export type MeQuery = { __typename?: 'Query', me?: { __typename?: 'User', id: string, email?: string | null } | null };
-
-
-export const LoginDocument = gql`
-    mutation login($email: String!, $password: String!) {
-  login(email: $email, password: $password) {
-    token
-  }
+export interface Price {
+  id: string;
+  product_id?: string;
+  active?: boolean;
+  description?: string;
+  unit_amount?: number;
+  currency?: string;
+  type?: Stripe.Price.Type;
+  interval?: Stripe.Price.Recurring.Interval;
+  interval_count?: number;
+  trial_period_days?: number | null;
+  metadata?: Stripe.Metadata;
+  products?: Product;
 }
-    `;
-export type LoginMutationFn = Apollo.MutationFunction<LoginMutation, LoginMutationVariables>;
 
-/**
- * __useLoginMutation__
- *
- * To run a mutation, you first call `useLoginMutation` within a React component and pass it any options that fit your needs.
- * When your component renders, `useLoginMutation` returns a tuple that includes:
- * - A mutate function that you can call at any time to execute the mutation
- * - An object with fields that represent the current status of the mutation's execution
- *
- * @param baseOptions options that will be passed into the mutation, supported options are listed on: https://www.apollographql.com/docs/react/api/react-hooks/#options-2;
- *
- * @example
- * const [loginMutation, { data, loading, error }] = useLoginMutation({
- *   variables: {
- *      email: // value for 'email'
- *      password: // value for 'password'
- *   },
- * });
- */
-export function useLoginMutation(baseOptions?: Apollo.MutationHookOptions<LoginMutation, LoginMutationVariables>) {
-        const options = {...defaultOptions, ...baseOptions}
-        return Apollo.useMutation<LoginMutation, LoginMutationVariables>(LoginDocument, options);
-      }
-export type LoginMutationHookResult = ReturnType<typeof useLoginMutation>;
-export type LoginMutationResult = Apollo.MutationResult<LoginMutation>;
-export type LoginMutationOptions = Apollo.BaseMutationOptions<LoginMutation, LoginMutationVariables>;
-export const SignupDocument = gql`
-    mutation signup($data: SignupInput!) {
-  signup(data: $data) {
-    token
-    user {
-      id
-    }
-  }
+export interface Customer {
+  id: string;
+  stripe_customer_id?: string;
 }
-    `;
-export type SignupMutationFn = Apollo.MutationFunction<SignupMutation, SignupMutationVariables>;
 
-/**
- * __useSignupMutation__
- *
- * To run a mutation, you first call `useSignupMutation` within a React component and pass it any options that fit your needs.
- * When your component renders, `useSignupMutation` returns a tuple that includes:
- * - A mutate function that you can call at any time to execute the mutation
- * - An object with fields that represent the current status of the mutation's execution
- *
- * @param baseOptions options that will be passed into the mutation, supported options are listed on: https://www.apollographql.com/docs/react/api/react-hooks/#options-2;
- *
- * @example
- * const [signupMutation, { data, loading, error }] = useSignupMutation({
- *   variables: {
- *      data: // value for 'data'
- *   },
- * });
- */
-export function useSignupMutation(baseOptions?: Apollo.MutationHookOptions<SignupMutation, SignupMutationVariables>) {
-        const options = {...defaultOptions, ...baseOptions}
-        return Apollo.useMutation<SignupMutation, SignupMutationVariables>(SignupDocument, options);
-      }
-export type SignupMutationHookResult = ReturnType<typeof useSignupMutation>;
-export type SignupMutationResult = Apollo.MutationResult<SignupMutation>;
-export type SignupMutationOptions = Apollo.BaseMutationOptions<SignupMutation, SignupMutationVariables>;
-export const MeDocument = gql`
-    query me {
-  me {
-    id
-    email
-  }
+export interface UserDetails {
+  id: string;
+  first_name: string;
+  last_name: string;
+  full_name?: string;
+  avatar_url?: string;
+  billing_address?: Stripe.Address;
+  payment_method?: Stripe.PaymentMethod[Stripe.PaymentMethod.Type];
 }
-    `;
 
-/**
- * __useMeQuery__
- *
- * To run a query within a React component, call `useMeQuery` and pass it any options that fit your needs.
- * When your component renders, `useMeQuery` returns an object from Apollo Client that contains loading, error, and data properties
- * you can use to render your UI.
- *
- * @param baseOptions options that will be passed into the query, supported options are listed on: https://www.apollographql.com/docs/react/api/react-hooks/#options;
- *
- * @example
- * const { data, loading, error } = useMeQuery({
- *   variables: {
- *   },
- * });
- */
-export function useMeQuery(baseOptions?: Apollo.QueryHookOptions<MeQuery, MeQueryVariables>) {
-        const options = {...defaultOptions, ...baseOptions}
-        return Apollo.useQuery<MeQuery, MeQueryVariables>(MeDocument, options);
-      }
-export function useMeLazyQuery(baseOptions?: Apollo.LazyQueryHookOptions<MeQuery, MeQueryVariables>) {
-          const options = {...defaultOptions, ...baseOptions}
-          return Apollo.useLazyQuery<MeQuery, MeQueryVariables>(MeDocument, options);
-        }
-export type MeQueryHookResult = ReturnType<typeof useMeQuery>;
-export type MeLazyQueryHookResult = ReturnType<typeof useMeLazyQuery>;
-export type MeQueryResult = Apollo.QueryResult<MeQuery, MeQueryVariables>;
+export interface ProductWithPrice extends Product {
+  prices?: Price[];
+}
+
+export interface Subscription {
+  id: string;
+  user_id: string;
+  status?: Stripe.Subscription.Status;
+  metadata?: Stripe.Metadata;
+  price_id?: string;
+  quantity?: number;
+  cancel_at_period_end?: boolean;
+  created: string;
+  current_period_start: string;
+  current_period_end: string;
+  ended_at?: string;
+  cancel_at?: string;
+  canceled_at?: string;
+  trial_start?: string;
+  trial_end?: string;
+  prices?: Price;
+}
